+++ views/layout_2-improved.html
@@ -1,5 +1,5 @@
 {{#if article.id}}
-<article class="article article--default">
+<article class="article article--default" data-capi-id="{{article.id}}">
 	<div class="article__colour-brand">
 		<div class="o-grid-row">
 			<div data-o-grid-colspan="12 XL2">
